refactor: split translation into two steps
Having a separate step for parsing all the repositories opens the door
to build logic which parses the repositories to find their names,
packages, etc. for auto-configuring which packages should be
available (we can't configure the build based on the output of a build
step, as far as I know).

This also makes the translation logic less messy, with less context
needed to pass around (and no more recursion).