fix: handle nullable types correctly
Closes #36

This commit moves the final determination of whether a type is nullable
to the type translation layer. This simplifies the translation of
higher-level elements, which no longer have to try to determine whether
nullability is appropriate, and allows better decisioning of whether a
type should actually be allowed to be nullable.

Additionally, this ends up solving several cases where array types
should have been translated as nullable but were not.