feat: add GIR fixup scripts
Works around #37 and potential future issues of the same kind

This commit adds the ability to run a "fixup script" on the input system
GIR files, which is expected to output any necessary overrides (created
via patching, etc.) to a target directory. The codegen process will use
these fixed up files in preference to its own.