fix: work around bad GObject-2.0 GIR
Works around #37
Closes #43

This commit adds a mechanism to the codegen step of the build to apply
XSLT stylesheets to broken GIR files using
[zig-build-libxslt](https://github.com/ianprime0509/zig-build-libxslt)
to build and run `xsltproc` on demand, without introducing an additional
system command dependency. Hopefully, this mechanism will be robust
enough to address any future upstream GIR issues without ugly hacks in
the codegen or waiting indefinitely for upstream.