build: always apply gir-fixes in codegen
Previously, even the common gir-fixes were only applied when a gir-profile was provided. This is inconvenient when running adhoc codegen for specific modules (e.g. for quick testing of just GTK functionality). Now, the common gir-fixes are applied even when no gir-profile is provided, but profile-specific fixes are still tied to their corresponding profile.