chore: disable Vulkan codegen and dependents
See #89. There is really no good way to address this right now without manually writing fixes or bindings for the 1000+ types in the Vulkan-1.0 GIR, which I don't want to do. A better solution will need to be identified in the long run, such as figuring out how to offload the Vulkan-1.0 bindings to a real set of Zig Vulkan bindings.